% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gt.R
\name{add_gt_grand_summary_row_mean}
\alias{add_gt_grand_summary_row_mean}
\title{Add a mean grand summary row to a gt table}
\usage{
add_gt_grand_summary_row_mean(gt_table, selection_type = "contains", labels)
}
\arguments{
\item{gt_table}{A \code{gt} table object.}

\item{selection_type}{Character string defining the selection method passed to \code{tidyselect_gt_column_name()}. Defaults to \code{"contains"}.}

\item{labels}{A configuration list generated by \code{make_gt_labels()} that defines which columns to summarize and ignore.}
}
\value{
A modified \code{gt} table with a grand summary "Mean" row.
}
\description{
Adds a "Mean" grand summary row to selected columns of a \code{gt} table using configuration from \code{make_gt_config()}.
}
\examples{
\dontrun{
labels <- make_gt_labels()
gt_tbl <- gt::gt(head(mtcars))
add_gt_grand_summary_row_mean(gt_tbl, labels = labels)
}
}
