% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gt_config.R
\name{make_gt_config}
\alias{make_gt_config}
\title{Make a gt configuration list}
\usage{
make_gt_config()
}
\value{
A list containing gt configuration settings.
}
\description{
Create a configuration list to control gt table styling (text, font, colors)
}
