% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gt_labels.R
\name{make_gt_labels}
\alias{make_gt_labels}
\title{Make a gt label list}
\usage{
make_gt_labels(
  metric_name = NULL,
  metric_cols = NULL,
  group_name = NULL,
  ignore_cols = NULL,
  format_cols_number = NULL,
  format_cols_integer = NULL,
  format_cols_percent = NULL,
  decimals_number = NULL,
  decimals_integer = NULL,
  decimals_percent = NULL,
  summarize_cols_total = NULL,
  summarize_cols_mean = NULL
)
}
\arguments{
\item{metric_name}{Character metric root name (e.g., effort, richness, occupancy).}

\item{metric_cols}{Character metric columns.}

\item{group_name}{Character name of grouping column (e.g., year, season).}

\item{ignore_cols}{Character vector of columns to exclude from processing.}

\item{format_cols_number}{Character vector of columns to format as numeric (e.g., mean).}

\item{format_cols_integer}{Character vector of columns to format as integer (e.g., minimum, maximum).}

\item{format_cols_percent}{Character vector of columns to format as percentage (e.g., percent).}

\item{decimals_number}{Numeric vector of number of decimals to use for numeric values.}

\item{decimals_integer}{Numeric vector of number of decimals to use for integer values.}

\item{decimals_percent}{Numeric vector of number of decimals to use for percent values.}

\item{summarize_cols_total}{Character vector of columns to summarize in total rows.}

\item{summarize_cols_mean}{Character vector of columns to summarize in mean rows.}
}
\value{
A list containing gt configuration settings.
}
\description{
Create a list to control column summarization. Defines names for metric and group. Defines names of columns by format (number, integer, percent), and those to include in summary rows for total and mean. Can also define columns to ignore, which will result in an empty cell (i.e., to replace "---"). Default column names that are accepted are mean, total, count, minimum, maximum, percent, and empty.
}
