% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_gt.R
\name{style_gt_diet}
\alias{style_gt_diet}
\title{Apply standard styling for diet (carnivore, omnivore, herbivore) gt tables}
\usage{
style_gt_diet(gt_table)
}
\arguments{
\item{gt_table}{A \code{gt} table object.}
}
\value{
A \code{gt} table styled for diet type.
}
\description{
Applies consistent colors and font styles for summary tables with diet type.
}
