% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_gt.R
\name{style_gt_images}
\alias{style_gt_images}
\title{Apply standard styling for image count gt tables}
\usage{
style_gt_images(gt_table)
}
\arguments{
\item{gt_table}{A \code{gt} table object.}
}
\value{
A \code{gt} table styled for image count.
}
\description{
Applies consistent font styles and column widths for summary tables with image count.
}
