% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_gt.R
\name{style_gt_zeros}
\alias{style_gt_zeros}
\title{Apply lighter styling to zeros for gt tables}
\usage{
style_gt_zeros(gt_table)
}
\arguments{
\item{gt_table}{A \code{gt} table object.}
}
\value{
A \code{gt} table styled for zero values.
}
\description{
Apply lighter styling to zeros for gt tables
}
