% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_gt_labels}
\alias{update_gt_labels}
\title{Update default column labels}
\usage{
update_gt_labels(new_labels, labels = NULL)
}
\arguments{
\item{new_labels}{A named list of new labels to apply to gt_labels.}

\item{labels}{A list containing keys for column name patterns.
Default values are defined in make_gt_labels()}
}
\value{
A modified \code{gt_labels} list.
}
\description{
Automatically formats numeric columns in a \code{gt} table based on patterns
defined in a configuration list.
}
\details{
This function revises the existing gt_labels list.
}
